+++ packages/ui/VirtualList/VirtualListBase.js
@@ -1054,9 +1054,7 @@ const VirtualListBase = (props) => {
 	delete rest.pageScroll;
 	delete rest.rtl;
 	delete rest.setChildAdapter;
-	delete rest.setUiChildAdapter;
 	delete rest.spacing;
-	delete rest.uiChildAdapter;
 	delete rest.updateStatesAndBounds;
 
 	if (variables.current.primary) {
